+++ RepairsApi/V2/Services/ICurrentUserLoader.cs
@@ -1,4 +1,4 @@
-﻿namespace RepairsApi.V2.MiddleWare
+﻿namespace RepairsApi.V2.Services
 {
     public interface ICurrentUserLoader
     {
